requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
google-analytics-data>=0.18.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.2
importlib-metadata>=6.0


